Last update before publish

+++ MyProfile.aspx.cs
@@ -25,7 +25,7 @@ namespace FinalProyect_MaxiPrograma_LVL3
                     Session.Add("Error", "You don't have permission to modify this profile. Please Loguin with User(no admin) to access this page.");
                     Response.Redirect("Error.aspx");
                 }
-                imgNewImage.ImageUrl = user.UrlImagen;
+                imgNewImage.ImageUrl = user.UrlImagen +"?v=" + DateTime.Now.Ticks.ToString();
                 if (user.UserName == null)
                     txtName.Text = "";
                 else
@@ -43,45 +43,57 @@ namespace FinalProyect_MaxiPrograma_LVL3
             user.UserName = txtName.Text;
             user.UserSurname = txtSurname.Text;
             string pass = txtOldPass.Text;
-            if (pass == "")
+            lblNewPass.Visible = false;
+            lblNewImage.Visible = false;
+            lblPass.Visible = false;
+            if (pass == user.Password)
             {
-                pass = user.Password;
-            }
-            else if (pass == user.Password)
-            {
-                if (txtPassword.Text == txtConfirmPassword.Text)
+                if (txtPassword.Text != txtConfirmPassword.Text)
                 {
-                    pass = txtPassword.Text;
+                    lblPass.Visible = false;
+                    lblNewPass.Visible = true;
+                    lblNewPass.Text = "Password don't match.";
                 }
                 else
                 {
-                    Session.Add("Error", "Password didn't match");
-                    Response.Redirect("Error.aspx");
+                    pass = txtPassword.Text;
+                    pass = txtPassword.Text;
+                    if (fupImage.HasFile)
+                    {
+                        // Verifica que el archivo sea una imagen
+                        if (!fupImage.FileName.EndsWith(".jpg") && !fupImage.FileName.EndsWith(".jpeg") && !fupImage.FileName.EndsWith(".png"))
+                        {
+                            lblImagenError.Visible = true;
+                            lblImagenError.Text = "Please upload a valid image";
+                            return;
+                        }
+                        string imagePath = Server.MapPath("~/Images/");
+                        string fileName = "profile-" + user.Id + ".png";
+                        if (File.Exists(imagePath + fileName))
+                        {
+                            File.Delete(imagePath + fileName);
+                        }
+                        fupImage.SaveAs(imagePath + fileName);
+                        Response.Cache.SetCacheability(HttpCacheability.NoCache);
+                        Response.Cache.SetExpires(DateTime.Now.AddSeconds(-1));
+                        Response.Cache.SetNoStore();
+                    }
+
+                    if (user.UrlImagen == null)
+                        negocio.correctImageUrl(user.Id);
+                    user.UserName = txtName.Text;
+                    user.UserSurname = txtSurname.Text;
+                    negocio.modify(user);
+                    Session["user"] = user;
+                    Response.Redirect("Default.aspx", false);
                 }
             }
             else
             {
-                Session.Add("Error", "Incorrect  Pasword");
-                Response.Redirect("Error.aspx");
+                lblNewPass.Visible = false;
+                lblPass.Visible = true;
+                lblPass.Text = "Invalid Password.";
             }
-            if (fupImage.HasFile)
-            {
-                string imagePath = Server.MapPath("~/Images/");
-                string fileName = "profile-" + user.Id + Path.GetExtension(fupImage.FileName);
-                if (File.Exists(imagePath + fileName))
-                {
-                    File.Delete(imagePath + fileName);
-                }
-                fupImage.SaveAs(imagePath + fileName);
-            }
-            if (user.UrlImagen == null)
-                negocio.correctImageUrl(user.Id);
-            user.UserName = txtName.Text;
-            user.UserSurname = txtSurname.Text;
-            negocio.modify(user);
-            Session["user"] = user;
-            Response.Redirect("Default.aspx", false);
-
         }
     }
 }
